# openapi.storefront.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commerce Platform Storefront API",
    "description": "Theme-facing page models and cart (spec §8.2). Reached through the edge only.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/storefront/v1/orders/{token}": {
      "get": {
        "tags": [
          "storefront"
        ],
        "operationId": "get_order",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "token",
            "in": "path",
            "description": "Order capability token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/orders/{token}/payment": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "The payment status (A10), polled by the order page while a payment is under way.",
        "operationId": "get_payment",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "token",
            "in": "path",
            "description": "Order capability token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/orders/{token}/payment-attempts": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "A new payment attempt after a failed one, within the order's payment window (A10).",
        "operationId": "new_attempt",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "token",
            "in": "path",
            "description": "Order capability token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStart"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "retry_not_allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/orders/{token}/payment-attempts/{attempt}/init": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Initializes (again) a pending attempt at its provider (A10): idempotent, for when the\nautomatic init after placement failed.",
        "operationId": "init_attempt",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "token",
            "in": "path",
            "description": "Order capability token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStart"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "attempt_not_pending",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptStatus": {
        "type": "string",
        "enum": [
          "pending",
          "succeeded",
          "failed",
          "expired"
        ]
      },
      "AttemptView": {
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          }
        }
      },
...
      "Carrier": {
        "type": "string",
        "enum": [
          "packeta_pickup",
          "packeta_home",
          "ppl",
          "personal_pickup"
        ]
      },
...
      "ChargeKind": {
        "type": "string",
        "enum": [
          "shipping",
          "payment_fee",
          "rounding"
        ]
      },
...
      "Currency": {
        "type": "string",
        "description": "Currencies used by EU markets: the euro and the member states' own currencies.\n`BGN` is historic (Bulgaria adopted the euro on 2026-01-01) and kept for old documents.",
        "enum": [
          "BGN",
          "CZK",
          "DKK",
          "EUR",
          "HUF",
          "PLN",
          "RON",
          "SEK"
        ]
      },
...
      "FulfillmentStatus": {
        "type": "string",
        "enum": [
          "unfulfilled",
          "label_created",
          "shipped",
          "delivered",
          "returned"
        ]
      },
...
      "MethodKind": {
        "type": "string",
        "enum": [
          "stripe",
          "bank_transfer",
          "cod",
          "fake"
        ]
      },
      "MoneyView": {
        "type": "object",
        "description": "The API shape of an amount (spec §8.1): `{amount_minor, currency, formatted}`.",
        "required": [
          "amount_minor",
          "currency",
          "formatted"
        ],
        "properties": {
          "amount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "formatted": {
            "type": "string",
            "example": "1 290,00 Kč"
          }
        }
      },
...
      "NextAction": {
        "oneOf": [
          {
            "type": "object",
            "description": "Go to the provider's page (relative or absolute URL).",
            "required": [
              "url",
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "redirect"
                ]
              },
              "url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Nothing to pay now (cash on delivery).",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "none"
                ]
              }
            }
          }
        ],
        "description": "What the customer does next."
      },
...
      "OrderAddress": {
        "type": "object",
        "required": [
          "name",
          "street",
          "city",
          "postal_code",
          "country"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "company": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "postal_code": {
            "type": "string"
          },
          "street": {
            "type": "string"
          }
        }
      },
      "OrderChargeView": {
        "type": "object",
        "required": [
          "kind",
          "total",
          "tax"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ChargeKind"
          },
          "tax": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyView"
          }
        }
      },
      "OrderLineView": {
        "type": "object",
        "required": [
          "sku",
          "name",
          "options_label",
          "quantity",
          "unit_price",
          "discount",
          "total",
          "tax_rate",
          "tax"
        ],
        "properties": {
          "discount": {
            "$ref": "#/components/schemas/MoneyView",
            "description": "This line's share of the coupon (A15)."
          },
          "name": {
            "type": "string"
          },
          "options_label": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "sku": {
            "type": "string"
          },
          "tax": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "tax_rate": {
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "unit_price": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "variant_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "processing",
          "shipped",
          "delivered",
          "cancelled",
          "returned"
        ]
      },
...
      "OrderView": {
        "type": "object",
        "description": "An order as its customer sees it (`/o/<token>`, the account).",
        "required": [
          "id",
          "number",
          "placed_at",
          "status",
          "fulfillment_status",
          "email",
          "locale",
          "currency",
          "lines",
          "charges",
          "subtotal",
          "discount",
          "shipping_total",
          "payment_fee",
          "rounding",
          "vat",
          "vat_total",
          "total",
          "shipping",
          "payment"
        ],
        "properties": {
          "billing_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OrderAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderChargeView"
            }
          },
          "coupon_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "discount": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "email": {
            "type": "string"
          },
          "exception": {
            "type": [
              "string",
              "null"
            ],
            "description": "`late_payment` (A10) when money arrived for an expired or cancelled order."
          },
          "fulfillment_status": {
            "$ref": "#/components/schemas/FulfillmentStatus"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineView"
            }
          },
          "locale": {
            "type": "string"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "number": {
            "type": "string"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentView"
          },
          "payment_fee": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "placed_at": {
            "type": "string",
            "format": "date-time"
          },
          "rounding": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "shipping": {
            "$ref": "#/components/schemas/ShippingSummary"
          },
          "shipping_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OrderAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "shipping_total": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "subtotal": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "vat": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VatRow"
            }
          },
          "vat_total": {
            "$ref": "#/components/schemas/MoneyView"
          }
        }
      },
...
      "PaymentStart": {
        "type": "object",
        "description": "What the client does after placing the order or starting a payment attempt.",
        "required": [
          "attempt_id"
        ],
        "properties": {
          "action": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NextAction",
                "description": "`null` when the provider could not be reached yet: retry with\n`POST /orders/{token}/payment-attempts/{attempt}/init` (A10)."
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "unpaid",
          "authorized",
          "paid",
          "partially_refunded",
          "refunded",
          "failed",
          "expired"
        ]
      },
      "PaymentView": {
        "type": "object",
        "required": [
          "method",
          "status",
          "can_retry"
        ],
        "properties": {
          "attempt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AttemptView",
                "description": "The latest attempt."
              },
              {
                "type": "null"
              }
            ]
          },
          "can_retry": {
            "type": "boolean",
            "description": "Whether the customer can start a new attempt now (A10)."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Unpaid orders are cancelled after this."
          },
          "method": {
            "$ref": "#/components/schemas/MethodKind"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          }
        }
      },
      "PickupPoint": {
        "type": "object",
        "description": "A pickup point as chosen in the carrier's widget (snapshot).",
        "required": [
          "id",
          "name",
          "street",
          "city",
          "zip",
          "country"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2, upper case."
          },
          "id": {
            "type": "string",
            "description": "The carrier's point id."
          },
          "name": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. `not_found`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`; `code` carries the specific error kind."
          }
        }
      },
...
      "ShippingSummary": {
        "type": "object",
        "required": [
          "carrier",
          "name"
        ],
        "properties": {
          "carrier": {
            "$ref": "#/components/schemas/Carrier"
          },
          "name": {
            "type": "string",
            "description": "The method's name in the order's locale, as it was at placement."
          },
          "pickup_point": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PickupPoint"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "VatRow": {
        "type": "object",
        "required": [
          "rate",
          "net",
          "vat",
          "gross"
        ],
        "properties": {
          "gross": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "net": {
            "$ref": "#/components/schemas/MoneyView"
          },
          "rate": {
            "type": "string",
            "description": "Percent (`\"21\"`)."
          },
          "vat": {
            "$ref": "#/components/schemas/MoneyView"
          }
        }
      }
    }
  },
...
}